requirements.txt
django==3.2.3
flake8
black
mypy
django-debug-toolbar==3.2.1
djangorestframework-jwt
djangorestframework
coverage==5.5
psycopg2-binary==2.9.1